오늘의 프로필 2개 소개 작업중 by laptop

+++ src/main/java/com/example/self_board_project/board/vo/Board.java
@@ -4,13 +4,12 @@ import com.example.self_board_project.core.support.Pagination;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.util.Date;
 @Getter
 @Setter
 public class Board extends Pagination {
     private Integer id;
-    private BigInteger userId;
+    private Integer userId;
     private String title;
     private String content;
     private Date createDate;
